As used for Web Directions South 2011 Exhibition. Very stable. As used for Web Directions South 2011 Exhibition. Very stable.
Doesn't crash, doesn't have any memory leaks based on my observation.
Have left this version running for over 2+ hours with no crash and no
loss of framerate.

One problem which is easily resolved yet I haven't done is when someone
zooms out to maximum, they can't zoom back in again. This is easily
rectified by a triple-tap to return to default-view, but is still a bug
which should be fixed.

UPDATES:
* double-tap is now triple-tap to return to default view
* to create the attractor, the user can now simply hold 5 fingers for a
second instead of scrunching. One finger will move the attractor; when
they are ready to make a planet, holding 5 fingers will create the
planet. The scrunch gesture also sort of still works, so if you scrunch
in/out, it will also work. This was modified because I realised during
numerous user observations that the scrunch gesture was too difficult
for many to pull off. This is mostly because of how it is currently
implemented: basically if one finger accidentally leaves the table for
a second during the scrunch-gesture, it won't register. And this is a
pretty easy thing to happen on the big multitouch table. When changing
it to a simple hold, and instructing the user to 'pull-off' quickly
once the planet was made, the illusion of a 'bang'-like force when the
planet is made was still delightful to them. This definitely needs more
tweaking though (for example, its currently possible to hold 5 fingers
down constantly and make a quick sequence of planets, which isn't
right.)
* there are two slightly different variations of textured planets now,
and which variation the planet will take is chosen at random when the
user creates one. One planet acts as they did when we finished
Information Visualisation Studio. The other variation looks more
music-visualisation-y: it responds clearly to sounds but somewhat
retains its shape. This variation also has a killer implosion-sequence
once they die; look out for it! The other variation currently has no
death sequence, it just disappears.
* A three-finger rub on a planet now works correctly, including
wireframe planets AND textured planets. Using one finger in this view
allows the user to rotate the planet around. Very cool to see the
implosion in this view. To get out of this view a triple-tap is
required.
* When the attractor is enabled, a pinch/zoom gesture on the attractor
will increase/decrease the radius of attraction. This is a great
inclusion because it allows a user to trap every single particle in the
space by expanding it to a huge size. Users had a lot of fun with this
feature!
* The borders for how far the particles can move out-of-bounds before
they appear back in the white-black-hole have been greatly reduced. A
keen observer might be able to see the border when grabbing a bunch of
particles and moving the attractor to a boundary, but keeping the
borders small gives a great impression of even more particles than
there really are: the white-black-hole looks like its continuously
spitting out new particles. Also, even if the boundary is found, its
pretty fun grabbing heaps of particles, and pushing them into the
boundary to be flung out of the white-black-hole at great speed.
* the planets now rotate at differing speeds